Refactor transaction components
